track loxodrome amm dexs and fees
